Issue #48: New ApiConfig option to allow sending of null values by default
Benefit: In the previous state of the library, it was hard to clear existing values because null values were ignored (e.g., clearing a date field never worked).

Downside: Moves some responsibility to think about which values should be nullable and which shouldn't to the user of the library - e.g., Salesforce seems to reject boolean values when they are sent as null. So now the library user needs to take care of ignoring certain null values with Jackson (`@JsonInclude(JsonInclude.Include.NON_NULL)`)